requirements.txt
streamlit
pandas
openpyxl
Pillow
xlsxwriter
folium
streamlit-folium
